requirements.txt
aniso8601==7.0.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-restql==0.10.2
django-safedelete==0.5.5
django-suit==0.2.28
djangorestframework==3.9.4
graphene==2.1.8
graphene-django==2.12.1
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
promise==2.3
pycodestyle==2.5.0
pycparser==2.20
pylint==2.3.1
pyPEG2==2.15.2
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
wrapt==1.11.2
whitenoise==5.2.0
